Crate nitro

Reexports

pub extern crate ncollide;

pub extern crate nphysics2d;

pub extern crate nalgebra;

Modules

audio

component

graphics

input

math

Structs

App
This structure represents a game instance.

Camera
Describes a viewpoint into the world.

GameObject
GameObjects are the primary descriptor for things that are visible in the game world. Players, items, enemies, and whatever else exists in your game world are typically GameObjects.

Enums

OptionLoaned
Special variant of Option that indicates a value may exist but is currently loaned out in a &mut.