Enum nitro::graphics::Sprite

pub enum Sprite {
    Texture(Texture),
    SpriteSheet(SpriteSheet),
}

Something which can be displayed on screen. This can optionally be animated.

Variants
Texture(Texture)

A static image. No animation is available.

SpriteSheet(SpriteSheet)

An animation given by a SpriteSheet.

Animation is not done automatically by Nitro, advancing through and selecting frames is typically the responsiblity of a Component delegated to controlling animation.