Struct nitro::graphics::SpriteSheet

pub struct SpriteSheet {
    pub animations: Vec<Vec<SpriteSheetFrame>>,
    pub current_animation: u32,
    pub current_frame: u32,
    // some fields omitted
}

An animation given by a SpriteSheet.

Animation is not done automatically by Nitro, advancing through and selecting frames is typically the responsiblity of a Component delegated to controlling animation.

Fields
animations: Vec<Vec<SpriteSheetFrame>> current_animation: u32 current_frame: u32