Module nitro::input

Modules

keyboard

mouse

Structs

Axis
An input axis, could be bound to W and S, A and D, up and down, or left and right.

Input
Allows you to query user input, and save and load keybindings and axes.

Enums

Button
A digital input, can be made to behave like an analog axis using nitro::input::Axis.