Struct nitro::math::Transform

pub struct Transform { /* fields omitted */ }

Methods

impl Transform

fn new(position: Vector, rotation: f32) -> Transform

fn from_x_y_rot(x: f32, y: f32, rotation: f32) -> Transform

Constructs a new transform from the given x position, y position, and rotation in radians.

fn position(&self) -> &Vector

fn mut_position(&mut self) -> &mut Vector

fn x(&self) -> &f32

X component of the position.

fn y(&self) -> &f32

Y component of the position.

fn rotation(&self) -> &f32

Rotation of this transform in radians

fn mut_x(&mut self) -> &mut f32

X component of the position.

fn mut_y(&mut self) -> &mut f32

Y component of the position.

fn mut_rotation(&mut self) -> &mut f32

Rotation of this transform in radians

fn forward(&self) -> Vector

This vector points forward relative to the Transform's rotation

Useful for making an object go forward. Multiply by -1 to get backwards. This direction could also be called up. It's a matter of perspective.

fn right(&self) -> Vector

This vector points to the right relative to the Transform's rotation

Useful for making an object go to its right. Multiply by -1 to get left.